refactor: move xoreutils tests
